move removal of typename and id function to shared package and also type guard check function
