+++ packages/frinx-workflow-builder/src/components/workflow-definition-modal/workflow-definition-modal.tsx
@@ -1,11 +1,11 @@
 import React, { FC } from 'react';
-import { ClientWorkflow, Editor } from '@frinx/shared';
+import { ClientWorkflowWithTasks, Editor, removeGraphqlSpecsFromWorkflow } from '@frinx/shared';
 import { Modal, ModalBody, ModalCloseButton, ModalContent, ModalHeader, ModalOverlay } from '@chakra-ui/react';
 
 type Props = {
   isOpen: boolean;
   onClose: () => void;
-  workflow: ClientWorkflow;
+  workflow: ClientWorkflowWithTasks;
 };
 
 const WorkflowDefinitionModal: FC<Props> = ({ isOpen, onClose, workflow }) => {
@@ -16,7 +16,7 @@ const WorkflowDefinitionModal: FC<Props> = ({ isOpen, onClose, workflow }) => {
         <ModalHeader>Workflow definition</ModalHeader>
         <ModalCloseButton />
         <ModalBody>
-          <Editor language="json" defaultValue={JSON.stringify(workflow, null, 2)} />
+          <Editor language="json" defaultValue={JSON.stringify(removeGraphqlSpecsFromWorkflow(workflow), null, 2)} />
         </ModalBody>
       </ModalContent>
     </Modal>
